moved breakpoint to operators, so catalog knows about it
